Added support for starting points
Added support for starting points. This uses the starting position
flag, which is a little bigger and better visible than the flag-red.png
I had been planning on using.

I may paint the number on to show which player starts on that hex.

All standard terrain codes should now be supported.