Turning grid off. For now.
Turning grid off for now as its inconsistently painting when changing
maps. Not really a game changer in the long run as changing maps all
the time, one right after the other, probably isn't representative of
how Wesnoth-Canvas will be used. But, seeing as it's weird behavior at
least with Safari 5 I'm going to postpone this feature.