Added editorName
'name' is what the game uses to display the title of the hex.
'editorName' is what the editor uses to display the title of the hex.

If there's no 'name' then the title doesn't get displayed in the game
(embellishments, etc.) If there's no 'editorName' use the regular
'name'. If this ends up confusing then I'll add 'name' to all terrains
at a minimum.