If no select, no coords written
Small change to reflect if no select box on canvas then no hex coords
are written.

With more time I'd like to have it so that when you move the mouse
outside the boundaries it takes the nearest row/column, as it does
currently on the top. Until then, no select box gets drawn in that
scenario.